requirements.txt
asgiref==3.8.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
idna==3.10
mock==5.1.0
mysqlclient==2.2.5
pillow==11.0.0
razorpay==1.4.2
reportlab==4.2.5
requests==2.32.3
setuptools==75.4.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
